Removes debug and test API endpoints
Removes the debug token info and test environment API endpoints.
These were likely used for development and are no longer needed in the production environment.
Also removes the 'approval_prompt' setting in the Google authentication configuration, which was forcing a refresh token on every login. This is no longer needed.